+++ python/CMakeLists.txt
@@ -109,8 +109,8 @@ python_module(TARGET_NAME revng-python-cli-commands MODULE_FILES
               ${REVNG_CLI_COMMANDS_MODULE_FILES})
 
 set(REVNG_CLI_COMMANDS_PTML_MODULE_FILES
-    revng/cli/_commands/ptml/cat.py revng/cli/_commands/ptml/common.py
-    revng/cli/_commands/ptml/__init__.py revng/cli/_commands/ptml/strip.py)
+    revng/cli/_commands/ptml/text.py revng/cli/_commands/ptml/common.py
+    revng/cli/_commands/ptml/__init__.py)
 python_module(TARGET_NAME revng-python-cli-commands-ptml MODULE_FILES
               ${REVNG_CLI_COMMANDS_PTML_MODULE_FILES})
